+++ app/js/helpers.js
@@ -1,0 +1,3 @@
+
+export const exists = (arr, f) =>
+  arr.reduce((acc, el) => f(el) | acc, false)
