+++ app/src/rooms/templates/rooms/index.html
@@ -5,26 +5,6 @@
 
 {% block content %}
 <div id="react-root">
-  <div class="section flexbox center-items-horizontal full-height">
-    <div class="container">
-      <div class="row">
-        {% for r in rooms %}
-        <div class="col s6 m4">
-          <div class="card blue-grey darken-1">
-            <div class="card-content white-text">
-              <span class="card-title">{{ r }}</span>
-              <p>{{ r.description }}</p>
-              <p> ? / {{ r.capacity }}</p>
-            </div>
-          <div class="card-action">
-            <a />
-          </div>
-          </div>
-        </div>
-        {% endfor %}
-      </div>
-    </div>
-  </div>
 </div>
 {% endblock %}
